Fix gradle build and Zuul timeout on first request
