+++ auth/src/main/java/hive/auth/AuthApplication.java
@@ -1,15 +1,18 @@
 package hive.auth;
 
+import hive.common.security.JwtConfig;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.loadbalancer.LoadBalanced;
 import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
 import org.springframework.web.client.RestTemplate;
 
 @SpringBootApplication
 @EnableEurekaClient
+@Import(JwtConfig.class)
 public class AuthApplication {
 
   public static void main(String[] args) {
