--- build.gradle
@@ -1,7 +1,0 @@
-subprojects {
-    afterEvaluate {
-        repositories {
-            mavenCentral()
-        }
-    }
-}
