+++ common/build.gradle
@@ -1,9 +1,19 @@
+buildscript {
+    repositories {
+        mavenCentral()
+        maven { url 'https://repo.spring.io/milestone' }
+    }
+}
+
 plugins {
-    id 'org.springframework.boot' version '2.1.3.RELEASE'
+    id "io.spring.dependency-management" version "1.0.5.RELEASE"
     id 'java'
 }
 
-apply plugin: 'io.spring.dependency-management'
+ext {
+    springBootVersion = '2.1.3.RELEASE'
+    springCloudVersion = 'Greenwich.SR1'
+}
 
 group = 'hive'
 version = '0.0.1-SNAPSHOT'
@@ -14,22 +24,19 @@ repositories {
     maven { url 'https://repo.spring.io/milestone' }
 }
 
-ext {
-    set('springCloudVersion', 'Greenwich.SR1')
-}
-
 dependencies {
     implementation 'javax.xml.bind:jaxb-api:2.3.1'
     implementation 'javax.activation:activation:1.1.1'
     implementation 'org.glassfish.jaxb:jaxb-runtime:2.3.2'
 
-    implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'org.springframework.boot:spring-boot-starter'
     implementation 'org.springframework.cloud:spring-cloud-starter-netflix-eureka-client'
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
 }
 
 dependencyManagement {
     imports {
+        mavenBom "org.springframework.boot:spring-boot-dependencies:${springBootVersion}"
         mavenBom "org.springframework.cloud:spring-cloud-dependencies:${springCloudVersion}"
     }
 }
