--- common/src/main/java/hive/common/CommonApplication.java
@@ -1,12 +1,0 @@
-package hive.common;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.EnableEurekaClient;
-
-@SpringBootApplication
-@EnableEurekaClient
-public class CommonApplication {
-  public static void main(String[] args) {
-    SpringApplication.run(CommonApplication.class, args);
-  }
-}
