+++ common/src/main/java/hive/common/entity/User.java
@@ -1,0 +1,60 @@
+package hive.common.entity;
+
+public class User {
+  private int id;
+  private String username;
+  private String password;
+  private String name;
+  private String role;
+
+  public User() {
+  }
+
+  public User(int id, String username, String password, String name, String role) {
+    this.id = id;
+    this.username = username;
+    this.password = password;
+    this.name = name;
+    this.role = role;
+  }
+
+  public int getId() {
+    return id;
+  }
+
+  public void setId(int id) {
+    this.id = id;
+  }
+
+  public String getUsername() {
+    return username;
+  }
+
+  public void setUsername(String username) {
+    this.username = username;
+  }
+
+  public String getPassword() {
+    return password;
+  }
+
+  public void setPassword(String password) {
+    this.password = password;
+  }
+
+  public String getName() {
+    return name;
+  }
+
+  public void setName(String name) {
+    this.name = name;
+  }
+
+  public String getRole() {
+    return role;
+  }
+
+  public void setRole(String role) {
+    this.role = role;
+  }
+}
