--- common/src/main/resources/application.properties
@@ -1,3 +1,0 @@
-spring.application.name=common-service
-server.port=9200
-eureka.client.service-url.default-zone=http://localhost:8761/eureka
