+++ gateway/src/main/java/hive/gateway/security/SecurityTokenConfig.java
@@ -27,21 +27,15 @@ public class SecurityTokenConfig extends WebSecurityConfigurerAdapter {
   protected void configure(HttpSecurity http) throws Exception {
     http
         .cors().and().csrf().disable()
-        // make sure we use stateless session; session won't be used to store user's state.
         .sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS)
         .and()
-        // handle an authorized attempts
         .exceptionHandling().authenticationEntryPoint((req, rsp, e) -> rsp.sendError(HttpServletResponse.SC_UNAUTHORIZED))
         .and()
-        // Add a filter to validate the tokens with every request
         .addFilterAfter(new JwtTokenAuthenticationFilter(jwtConfig), UsernamePasswordAuthenticationFilter.class)
-        // authorization requests config
         .authorizeRequests()
-        // allow all who are accessing "auth" service
         .antMatchers(HttpMethod.POST, jwtConfig.getUri()).permitAll()
-        // must be an admin if trying to access admin area (authentication is also required here)
+        // change this later
         .antMatchers("/gallery" + "/admin/**").hasRole("ADMIN")
-        // Any other request must be authenticated
         .anyRequest().authenticated();
   }
 
@@ -50,6 +44,8 @@ public class SecurityTokenConfig extends WebSecurityConfigurerAdapter {
     return new JwtConfig();
   }
 
+
+  // TODO change this to use properties
   @Bean
   public CorsConfigurationSource corsConfigurationSource() {
     final var config = new CorsConfiguration();
