+++ gateway/src/main/resources/application.properties
@@ -9,3 +9,6 @@ zuul.routes.auth-service.service-id=AUTH-SERVICE
 zuul.routes.auth-service.strip-prefix=false
 # Exclude authorization from sensitive headers ?
 zuul.routes.auth-service.sensitive-headers=Cookie,Set-Cookie
+
+# Fix timeout error on first request
+zuul.ribbon.eager-load.enabled=true
