+++ repository/build.gradle
@@ -1,6 +1,22 @@
+buildscript {
+    ext {
+        springBootVersion = '2.1.3.RELEASE'
+        springCloudVersion = 'Greenwich.SR1'
+    }
+    repositories {
+        mavenCentral()
+        maven { url 'https://repo.spring.io/milestone' }
+    }
+    dependencies {
+        classpath "org.springframework.boot:spring-boot-gradle-plugin:${springBootVersion}" as Object
+        classpath "org.springframework.cloud:spring-cloud-dependencies:${springCloudVersion}" as Object
+    }
+}
+
 plugins {
-    id 'org.springframework.boot' version '2.1.3.RELEASE'
+    id "io.spring.dependency-management" version "1.0.5.RELEASE"
     id 'java'
+    id 'org.springframework.boot' version '2.1.3.RELEASE'
 }
 
 apply plugin: 'io.spring.dependency-management'
@@ -12,10 +28,6 @@ sourceCompatibility = '11'
 repositories {
     mavenCentral()
     maven { url 'https://repo.spring.io/milestone' }
-}
-
-ext {
-    set('springCloudVersion', 'Greenwich.SR1')
 }
 
 dependencies {
@@ -35,6 +47,7 @@ dependencies {
 
 dependencyManagement {
     imports {
+        mavenBom "org.springframework.boot:spring-boot-dependencies:${springBootVersion}"
         mavenBom "org.springframework.cloud:spring-cloud-dependencies:${springCloudVersion}"
     }
 }
